refactor(core): consolidate scattered functions into Zoxide class (#3)
